+++ CHANGELOG.md
@@ -1,7 +1,29 @@
 # Change Log
 
 
-## [0.0.20/0.0.21]
+## [0.0.22]
+- [**Feature**] Restart dothttp-cli command
+- [**Improvement**] save file before running (for http file, for notebook it will not try to save)
+- [**Bug**] import postman shows error message in case of failure
+- [**Feature**] supports vscode hovers(shows informational message on hovering in dothttp)
+- [**Feature**] supports vscode definition(on clicking name, it will direct you to base definition)
+- [**Improvement**] postman import with file(earlier it used to be link, now supports import via locally downloaded file)
+- [**Improvement**] Bump dothttp-cli to 0.0.24
+  - [**Improvement**] import postman now supports 2.1.0 postman collection (also lot of bug fixes with import)
+
+## [0.0.21]
+- [**Improvement**] Bump dothttp-cli to 0.0.23
+  - [**Feature**] certificates pinning for sites
+    - P12 (with password)
+    - cert (with cert & optional key)
+  - [**Feature**] Allow insecure requests
+  - [**Feature**] Cleanup Session after request completion
+  - [**Improvement**] Extend url from base (base_http url + target_def url)
+  - [**Improvement**] Extend certificate, insecure, clear from base
+- [**Improvement**] expand url query to dothttp format
+
+
+## [0.0.20]
 - [**Feature**] release notebook changes
 - Sync Pull latest changes for notebook
 ## [0.0.19]